First draft of our model, version 0.1
This is a base draft of our machine learning model for our Sendy delivery logistics project. This version creates a linear relationship between two variables, the total distance between the pickup point and its corresponding destination (X), and the total time it takes to arrive from the respective points (Y). More features and optimization to follow.